% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{checkRepeatedDesign}
\alias{checkRepeatedDesign}
\title{Check if data represents repeated measurements design}
\usage{
checkRepeatedDesign(summarization_output)
}
\arguments{
\item{summarization_output}{output of the dataProcess function}
}
\value{
logical, TRUE if data represent repeated measurements design
}
\description{
Check if data represents repeated measurements design
}
