% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{countMultiRun}
\alias{countMultiRun}
\title{This counts how many runs are in a data set.}
\usage{
countMultiRun(data)
}
\description{
If there are multiple runs, then we have some normalization options available.
}
\details{
@param data  The data to count.
 @return  A list containing information whether this is risky or not.
}
