% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{decide_missingness}
\alias{decide_missingness}
\title{Examine the work data structure and figure out how much stuff is missing.}
\usage{
decide_missingness(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
