% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{DIANNtoMSstatsFormat}
\alias{DIANNtoMSstatsFormat}
\title{Import Spectronaut files}
\usage{
DIANNtoMSstatsFormat(
  diann_quantification,
  annotation,
  protein_id_column = "ProteinsIds",
  fragment_column = "FragmentQuantCorrected",
  useUniquePeptide = TRUE,
  summaryforMultipleRows = max,
  removeFewMeasurements = TRUE,
  removeProtein_with1Peptide = FALSE,
  filter_qvalue = TRUE,
  qvalue_cutoff = 0.01,
  filter_prot_qvalue = FALSE,
  prot_qvalue_cutoff = 0.01,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL,
  ...
)
}
\arguments{
\item{diann_quantification}{}

\item{annotation}{annotation tableincludes Condition, BioReplicate, Run.}

\item{protein_id_column}{name of a column that will be used as protein ID.}

\item{fragment_column}{name of a column that stores fragment intensities.}

\item{useUniquePeptide}{TRUE (default) removes peptides that are assigned for more than one proteins. 
We assume to use unique peptide for each protein.}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple measurements for certain feature and certain run, use highest or sum of multiple intensities.}

\item{removeFewMeasurements}{TRUE (default) will remove the features that have 1 or 2 measurements across runs.}

\item{removeProtein_with1Peptide}{TRUE will remove the proteins which have only 1 peptide and charge. FALSE is default.}

\item{filter_qvalue}{TRUE(default) will filter out the intensities that have greater than qvalue_cutoff in EG.Qvalue column. Those intensities will be replaced with zero and will be considered as censored missing values for imputation purpose.}

\item{qvalue_cutoff}{Cutoff for EG.Qvalue. default is 0.01.}

\item{filter_prot_qvalue}{TRUE(default) will filter out the intensities that have greater than qvalue_cutoff in EG.Qvalue column. Those intensities will be replaced with zero and will be considered as censored missing values for imputation purpose.}

\item{prot_qvalue_cutoff}{Cutoff for EG.Qvalue. default is 0.01.}

\item{use_log_file}{logical. If TRUE, information about data processing
will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be added
to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing wil be printed
to the console.}

\item{log_file_path}{character. Path to a file to which information about 
data processing will be saved. 
If not provided, such a file will be created automatically.
If `append = TRUE`, has to be a valid path to a file.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame in the MSstats required format.
}
\description{
Import Spectronaut files
}
\author{
Mateusz Staniak
}
