% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_converters.R
\name{DIAUmpiretoMSstatsFormat}
\alias{DIAUmpiretoMSstatsFormat}
\title{Import DIA-Umpire files}
\usage{
DIAUmpiretoMSstatsFormat(
  raw.frag,
  raw.pep,
  raw.pro,
  annotation,
  useSelectedFrag = TRUE,
  useSelectedPep = TRUE,
  fewMeasurements = "remove",
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  ...
)
}
\arguments{
\item{raw.frag}{name of FragSummary_date.xls data, which includes feature-level data.}

\item{raw.pep}{name of PeptideSummary_date.xls data, which includes selected fragments information.}

\item{raw.pro}{name of ProteinSummary_date.xls data, which includes selected peptides information.}

\item{annotation}{name of annotation data which includes Condition, BioReplicate, Run information.}

\item{useSelectedFrag}{TRUE will use the selected fragment for each peptide. 'Selected_fragments' column is required.}

\item{useSelectedPep}{TRUE will use the selected peptide for each protein. 'Selected_peptides' column is required.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import DIA-Umpire files
}
\author{
Meena Choi, Olga Vitek
}
