% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addCoverageInfo}
\alias{.addCoverageInfo}
\title{Add coverage information to a data.table}
\usage{
.addCoverageInfo(input)
}
\arguments{
\item{input}{data.table}
}
\value{
data.table
}
\description{
Add coverage information to a data.table
}
\keyword{internal}
