% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addModelVariances}
\alias{.addModelVariances}
\title{Add model variances}
\usage{
.addModelVariances(input)
}
\arguments{
\item{input}{data.table}
}
\value{
data.table
}
\description{
Add model variances
}
\keyword{internal}
