% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addNInformativeInfo}
\alias{.addNInformativeInfo}
\title{Add information about number of informative features}
\usage{
.addNInformativeInfo(input, min_feature_count)
}
\arguments{
\item{input}{data.table}

\item{min_feature_count}{minimum number of quality features to consider}
}
\value{
data.table
}
\description{
Add information about number of informative features
}
\keyword{internal}
