% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addNoisyFlag}
\alias{.addNoisyFlag}
\title{Add flag for noisy features}
\usage{
.addNoisyFlag(input)
}
\arguments{
\item{input}{data.table}
}
\value{
data.table
}
\description{
Add flag for noisy features
}
\keyword{internal}
