% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addOutlierCutoff}
\alias{.addOutlierCutoff}
\title{Add outlier cutoff}
\usage{
.addOutlierCutoff(input, quantile_order = 0.01)
}
\arguments{
\item{input}{data.table}

\item{quantile_order}{quantile used to label outliers}
}
\value{
data.table
}
\description{
Add outlier cutoff
}
\keyword{internal}
