% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addOutlierInformation}
\alias{.addOutlierInformation}
\title{Add flag for outlier}
\usage{
.addOutlierInformation(input, tol = 3, keep_run = FALSE)
}
\arguments{
\item{input}{data.table}

\item{tol}{cutoff for outliers}

\item{keep_run}{if TRUE, completely missing runs will be kept}
}
\value{
logical
}
\description{
Add flag for outlier
}
\keyword{internal}
