% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imputation.R
\name{.addSurvivalPredictions}
\alias{.addSurvivalPredictions}
\title{Get predicted values from a survival model}
\usage{
.addSurvivalPredictions(input)
}
\arguments{
\item{input}{data.table}
}
\value{
numeric vector of predictions
}
\description{
Get predicted values from a survival model
}
\keyword{internal}
