% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.adjustLRuns}
\alias{.adjustLRuns}
\title{Adjust summarized abundance based on the heavy channel}
\usage{
.adjustLRuns(input, rename = FALSE)
}
\arguments{
\item{input}{data.table}

\item{rename}{if TRUE, rename the output column to LogIntensities}
}
\value{
data.table
}
\description{
Adjust summarized abundance based on the heavy channel
}
\keyword{internal}
