% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.calculateOutlierCutoff}
\alias{.calculateOutlierCutoff}
\title{Calculate cutoff to label outliers}
\usage{
.calculateOutlierCutoff(input, quantile_order = 0.01)
}
\arguments{
\item{input}{data.table}

\item{quantile_order}{quantile used to label outliers}
}
\value{
numeric
}
\description{
Calculate cutoff to label outliers
}
\keyword{internal}
