% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_checks.R
\name{.checkContrastMatrix}
\alias{.checkContrastMatrix}
\title{Check if contrast matrix includes all conditions}
\usage{
.checkContrastMatrix(contrast_matrix, input)
}
\arguments{
\item{contrast_matrix}{contrast matrix}

\item{input}{data.table of summarized data}
}
\description{
Check if contrast matrix includes all conditions
}
\keyword{internal}
