% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.checkExperimentDesign}
\alias{.checkExperimentDesign}
\title{Check if a given column exists in the data}
\usage{
.checkExperimentDesign(input, column_name)
}
\arguments{
\item{input}{data.table}

\item{column_name}{chr, name of a column to check}
}
\description{
Check if a given column exists in the data
}
\keyword{internal}
