% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.checkGCPlotsInput}
\alias{.checkGCPlotsInput}
\title{Check groupComparisonPlots parameters}
\usage{
.checkGCPlotsInput(type, log_base, selected_labels, all_labels)
}
\arguments{
\item{type}{type of a plot: HEATMAP/VOLCANOPLOT/COMPARISONPLOT}

\item{log_base}{2 or 10}

\item{selected_labels}{character vector of contrast labels}

\item{all_labels}{character vector of all contrast labels}
}
\description{
Check groupComparisonPlots parameters
}
\keyword{internal}
