% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_checks.R
\name{.checkGroupComparisonInput}
\alias{.checkGroupComparisonInput}
\title{Check if groupComparison input was processed by the dataProcess function}
\usage{
.checkGroupComparisonInput(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Check if groupComparison input was processed by the dataProcess function
}
\keyword{internal}
