% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_checks.R
\name{.checkRepeated}
\alias{.checkRepeated}
\title{Check if the experiment consist of repeated measurements (time course)}
\usage{
.checkRepeated(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Check if the experiment consist of repeated measurements (time course)
}
\keyword{internal}
