% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.checkSingleFeature}
\alias{.checkSingleFeature}
\title{Check if data has less than two features}
\usage{
.checkSingleFeature(input)
}
\arguments{
\item{input}{data.table}
}
\value{
logical
}
\description{
Check if data has less than two features
}
\keyword{internal}
