% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{.checkSingleLabelProteins}
\alias{.checkSingleLabelProteins}
\title{Check if there are proteins with a single label in a labeled dataset}
\usage{
.checkSingleLabelProteins(input)
}
\arguments{
\item{input}{data.table}
}
\value{
TRUE invisibly
}
\description{
Check if there are proteins with a single label in a labeled dataset
}
\keyword{internal}
