% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_checks.R
\name{.checkSingleSubject}
\alias{.checkSingleSubject}
\title{Check if there is only single subject}
\usage{
.checkSingleSubject(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Check if there is only single subject
}
\keyword{internal}
