% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_checks.R
\name{.checkTechReplicate}
\alias{.checkTechReplicate}
\title{Check if there are technical replicates}
\usage{
.checkTechReplicate(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Check if there are technical replicates
}
\keyword{internal}
