% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.checkUnProcessedDataValidity}
\alias{.checkUnProcessedDataValidity}
\title{Check validity of data that were not processed by MSstats converter}
\usage{
.checkUnProcessedDataValidity(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Check validity of data that were not processed by MSstats converter
}
\keyword{internal}
