% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{.finalizeInput}
\alias{.finalizeInput}
\title{Add summary statistics to dataProcess output}
\usage{
.finalizeInput(input, predicted_survival, method, impute, censored_symbol)
}
\arguments{
\item{input}{feature-level data}

\item{method}{summary method}

\item{impute}{if TRUE, censored missing values were imputed}

\item{censored_symbol}{censored missing value indicator}

\item{summarized}{protein-level data (list)}
}
\description{
Add summary statistics to dataProcess output
}
\keyword{internal}
