% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{.finalizeLinear}
\alias{.finalizeLinear}
\title{Summary statistics for linear model-based summarization}
\usage{
.finalizeLinear(input, censored_symbol)
}
\arguments{
\item{input}{feature-level data}

\item{censored_symbol}{censored missing value indicator}
}
\description{
Summary statistics for linear model-based summarization
}
\keyword{internal}
