% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.fitHuber}
\alias{.fitHuber}
\title{Wrapper to fit robust linear model for one protein}
\usage{
.fitHuber(input)
}
\value{
rlm
}
\description{
Wrapper to fit robust linear model for one protein
}
\keyword{internal}
