% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.fitHuberLSQR}
\alias{.fitHuberLSQR}
\title{Wrapper to fit robust linear model for one protein with a matrix-free
LSQR-based solver instead of \code{MASS::rlm}'s QR-based one}
\usage{
.fitHuberLSQR(input)
}
\value{
list, see \code{.fitRobustRunFeatureModel}
}
\description{
Wrapper to fit robust linear model for one protein with a matrix-free
LSQR-based solver instead of \code{MASS::rlm}'s QR-based one
}
\keyword{internal}
