% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.fitLinearModel}
\alias{.fitLinearModel}
\title{Fit a linear model}
\usage{
.fitLinearModel(input, is_single_feature, is_labeled, equal_variances)
}
\arguments{
\item{input}{data.table}

\item{is_single_feature}{logical, if TRUE, data has single feature}

\item{is_labeled}{logical, if TRUE, data comes from a labeled experiment}

\item{equal_variances}{logical, if TRUE, equal variances are assumed}
}
\value{
lm or merMod
}
\description{
Fit a linear model
}
\keyword{internal}
