% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.flagLowCoverage}
\alias{.flagLowCoverage}
\title{Flag for low coverage features}
\usage{
.flagLowCoverage(input)
}
\arguments{
\item{input}{data.table}
}
\value{
logical
}
\description{
Flag for low coverage features
}
\keyword{internal}
