% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.getAllComparisons}
\alias{.getAllComparisons}
\title{Get all comparisons for a single protein and a contrast matrix}
\usage{
.getAllComparisons(input, fitted_model, contrast_matrix, groups, protein)
}
\arguments{
\item{input}{summarized data}

\item{fitted_model}{model fitted by the .fitModelForGroupComparison function}

\item{contrast_matrix}{contrast matrix}

\item{groups}{unique labels of experimental conditions}

\item{protein}{name of a protein}
}
\description{
Get all comparisons for a single protein and a contrast matrix
}
\keyword{internal}
