% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.getColorKeyGGPlot2}
\alias{.getColorKeyGGPlot2}
\title{Create colorkey for ggplot2 heatmap}
\usage{
.getColorKeyGGPlot2(my.colors, blocks)
}
\arguments{
\item{my.colors}{blocks}
}
\description{
Create colorkey for ggplot2 heatmap
}
\keyword{internal}
