% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.getColorKeyPlotly}
\alias{.getColorKeyPlotly}
\title{Create colorkey for plotly heatmap}
\usage{
.getColorKeyPlotly(my.colors, blocks)
}
\arguments{
\item{my.colors}{blocks}
}
\description{
Create colorkey for plotly heatmap
}
\keyword{internal}
