% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_contrasts.R
\name{.getContrastLabels}
\alias{.getContrastLabels}
\title{Get labels for contrasts}
\usage{
.getContrastLabels(contrasts)
}
\arguments{
\item{contrasts}{list of lists of condition labels}
}
\description{
Get labels for contrasts
}
\keyword{internal}
