% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.getEmptyComparison}
\alias{.getEmptyComparison}
\title{Comparison output when there are measurements only in a single condition}
\usage{
.getEmptyComparison(input, contrast_matrix, groups, protein)
}
\arguments{
\item{input}{summarized data}

\item{contrast_matrix}{contrast matrix}

\item{groups}{unique labels of experimental conditions}

\item{protein}{name of a protein}
}
\description{
Comparison output when there are measurements only in a single condition
}
\keyword{internal}
