% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.getFeatureVariances}
\alias{.getFeatureVariances}
\title{Calculate variances of features}
\usage{
.getFeatureVariances(input, tolerance = 3)
}
\arguments{
\item{input}{data.table}

\item{tolerance}{cutoff for outliers}
}
\value{
numeric
}
\description{
Calculate variances of features
}
\keyword{internal}
