% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_checks.R
\name{.getLogBaseName}
\alias{.getLogBaseName}
\title{Check if logarithm with base 2 or 10 was used}
\usage{
.getLogBaseName(processed)
}
\arguments{
\item{processed}{data.table}
}
\description{
Check if logarithm with base 2 or 10 was used
}
\keyword{internal}
