% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.getMedian}
\alias{.getMedian}
\title{Get median of protein abundances for a given label}
\usage{
.getMedian(df, label)
}
\arguments{
\item{df}{`data.table`}

\item{label}{"L" for light isotopes, "H" for heavy isotopes.}
}
\description{
Get median of protein abundances for a given label
}
\keyword{internal}
