% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSize.R
\name{.getMedianSigmaSubject}
\alias{.getMedianSigmaSubject}
\title{Get median per subject or group by subject}
\usage{
.getMedianSigmaSubject(var_component)
}
\arguments{
\item{var_component}{data.frame, output of .getVarComponent}
}
\description{
Get median per subject or group by subject
}
\keyword{internal}
