% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_censored.R
\name{.getMin}
\alias{.getMin}
\title{Utility function: get 0.99 * minimum of non-missing values}
\usage{
.getMin(abundance, nonmissing)
}
\arguments{
\item{abundance}{abundances values}

\item{nonmissing}{logical vector}
}
\description{
Utility function: get 0.99 * minimum of non-missing values
}
\keyword{internal}
