% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.getModelParameters}
\alias{.getModelParameters}
\title{Get params (coefficients, covariance matrix, degrees of freedom) from a model}
\usage{
.getModelParameters(fitted_model)
}
\arguments{
\item{fitted_model}{object of class lm or lmerMod}
}
\description{
Get params (coefficients, covariance matrix, degrees of freedom) from a model
}
\keyword{internal}
