% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_censored.R
\name{.getNonMissingFilter}
\alias{.getNonMissingFilter}
\title{Identify non-missing values}
\usage{
.getNonMissingFilter(input, impute, censored_symbol)
}
\arguments{
\item{input}{`data.table` in MSstats format}

\item{impute}{if TRUE, missing values are supposed to be imputed}

\item{censored_symbol}{`censoredInt` parameter to dataProcess}
}
\description{
Identify non-missing values
}
\keyword{internal}
