% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocess_plots.R
\name{.getSingleProteinForProfile}
\alias{.getSingleProteinForProfile}
\title{Get data for a single protein to plot}
\usage{
.getSingleProteinForProfile(processed, all_proteins, i)
}
\arguments{
\item{all_proteins}{character, set of protein names}

\item{i}{integer, index of protein to use}

\item{dataProcess}{output -> FeatureLevelData}
}
\description{
Get data for a single protein to plot
}
\keyword{internal}
