% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.getWideTable}
\alias{.getWideTable}
\title{Utility function for quantile normalization - get table in wide format}
\usage{
.getWideTable(input, runs, label = "L", remove_missing = TRUE)
}
\arguments{
\item{input}{`data.table` in MSstats standard format}

\item{label}{"L" for light isotopes, "H" for heavy isotopes}

\item{remove_missing}{if TRUE, only non-missing values will be considered}

\item{vector}{of run labels}
}
\description{
Utility function for quantile normalization - get table in wide format
}
\keyword{internal}
