% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.groupComparisonWithMultipleCores}
\alias{.groupComparisonWithMultipleCores}
\title{Perform group comparison per protein in parallel}
\usage{
.groupComparisonWithMultipleCores(
  summarized_list,
  contrast_matrix,
  weights,
  save_fitted_models,
  repeated,
  samples_info,
  numberOfCores
)
}
\arguments{
\item{summarized_list}{output of MSstatsPrepareForGroupComparison}

\item{contrast_matrix}{contrast matrix}

\item{save_fitted_models}{if TRUE, fitted models will be included in the output}

\item{repeated}{logical, output of checkRepeatedDesign function}

\item{samples_info}{data.table, output of getSamplesInfo function}

\item{numberOfCores}{Number of cores for parallel processing. 
A logfile named `MSstats_groupComparison_log_progress.log` is created to 
track progress. Only works for Linux & Mac OS.}
}
\description{
Perform group comparison per protein in parallel
}
\keyword{internal}
