% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.handleEmptyConditions}
\alias{.handleEmptyConditions}
\title{Handle contrast when some of the conditions are missing}
\usage{
.handleEmptyConditions(
  input,
  fit,
  contrast,
  groups,
  parameters,
  protein,
  empty_conditions,
  coefs
)
}
\arguments{
\item{input}{summarized data}

\item{contrast}{single row of a contrast matrix}

\item{groups}{unique labels of experimental conditions}

\item{parameters}{parameters extracted from the model}

\item{protein}{name of a protein}

\item{empty_conditions}{labels of empty conditions}

\item{coefs}{coefficient of the fitted model}
}
\description{
Handle contrast when some of the conditions are missing
}
\keyword{internal}
