% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.handleSingleContrast}
\alias{.handleSingleContrast}
\title{Group comparison for a single contrast}
\usage{
.handleSingleContrast(input, fit, contrast, groups, parameters, protein, coefs)
}
\arguments{
\item{input}{summarized data}

\item{contrast}{single row of a contrast matrix}

\item{groups}{unique labels of experimental conditions}

\item{parameters}{parameters extracted from the model}

\item{protein}{name of a protein}

\item{coefs}{coefficient of the fitted model}
}
\description{
Group comparison for a single contrast
}
\keyword{internal}
