% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{.logDatasetInformation}
\alias{.logDatasetInformation}
\title{Log information about feature-level data}
\usage{
.logDatasetInformation(input)
}
\arguments{
\item{input}{data.table}
}
\value{
TRUE invisibly after successful logging
}
\description{
Log information about feature-level data
}
\keyword{internal}
