% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{.logMissingness}
\alias{.logMissingness}
\title{Log information about missing data}
\usage{
.logMissingness(input)
}
\arguments{
\item{input}{data.table}
}
\value{
TRUE invisibly
}
\description{
Log information about missing data
}
\keyword{internal}
