% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{.logSingleLabeledProteins}
\alias{.logSingleLabeledProteins}
\title{Print proteins with a single label to the log file}
\usage{
.logSingleLabeledProteins(input, label)
}
\arguments{
\item{input}{data.table}

\item{label}{label ("L" or "H")}
}
\value{
TRUE invisibly
}
\description{
Print proteins with a single label to the log file
}
\keyword{internal}
