% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{.logSummaryStatistics}
\alias{.logSummaryStatistics}
\title{Print summary statistics to the log file}
\usage{
.logSummaryStatistics(input)
}
\arguments{
\item{input}{data.table}
}
\value{
TRUE invisibly
}
\description{
Print summary statistics to the log file
}
\keyword{internal}
