% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.makeFactorColumns}
\alias{.makeFactorColumns}
\title{Make factor columns where needed}
\usage{
.makeFactorColumns(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Make factor columns where needed
}
\keyword{internal}
