% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.makeHeatmap}
\alias{.makeHeatmap}
\title{Create heatmap}
\usage{
.makeHeatmap(input, my.colors, my.breaks, x.axis.size, y.axis.size)
}
\arguments{
\item{input}{data.table}

\item{x.axis.size}{size of axes labels, e.g. name of the comparisons in heatmap, and in comparison plot. Default is 10.}

\item{y.axis.size}{size of axes labels, e.g. name of targeted proteins in heatmap. Default is 10.}
}
\description{
Create heatmap
}
\keyword{internal}
