% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.makeHeatmapPlotly}
\alias{.makeHeatmapPlotly}
\title{Create heatmap}
\usage{
.makeHeatmapPlotly(
  input,
  my.colors,
  my.breaks,
  x.axis.size,
  y.axis.size,
  height,
  numProtein
)
}
\arguments{
\item{input}{data.table}

\item{x.axis.size}{size of axes labels, e.g. name of the comparisons in heatmap, and in comparison plot. Default is 10.}

\item{y.axis.size}{size of axes labels, e.g. name of targeted proteins in heatmap. Default is 10.}

\item{height}{height of the saved file in pixels. Default is 600.}

\item{numProtein}{For ggplot2: The number of proteins which will be presented in each heatmap. Default is 100. Maximum possible number of protein for one heatmap is 180.
For Plotly: use this parameter to adjust the number of proteins to be displayed on the heatmap}
}
\description{
Create heatmap
}
\keyword{internal}
