% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocess_plots.R
\name{.makeProfilePlot}
\alias{.makeProfilePlot}
\title{Create profile plot}
\usage{
.makeProfilePlot(
  input,
  is_censored,
  featureName,
  y.limdown,
  y.limup,
  x.axis.size,
  y.axis.size,
  text.size,
  text.angle,
  legend.size,
  dot.size.profile,
  ss,
  s,
  cumGroupAxis,
  yaxis.name,
  lineNameAxis,
  groupNametemp,
  dot_colors,
  condition.layout = NULL
)
}
\arguments{
\item{input}{data.table}

\item{is_censored}{TRUE if censored values were imputed}

\item{featureName}{for "ProfilePlot" only, "Transition" (default) means 
printing feature legend in transition-level; "Peptide" means printing feature 
legend in peptide-level; "NA" means no feature legend printing.}

\item{x.axis.size}{size of x-axis labeling for "Run" in Profile Plot and 
QC Plot, and "Condition" in Condition Plot. Default is 10.}

\item{y.axis.size}{size of y-axis labels. Default is 10.}

\item{text.size}{size of labels represented each condition at the top of 
graph in Profile Plot and QC plot. Default is 4.}

\item{text.angle}{angle of labels represented each condition at the top
of graph in Profile Plot and QC plot or x-axis labeling in Condition plot. 
Default is 0.}

\item{legend.size}{size of feature legend (transition-level or peptide-level)
above graph in Profile Plot. Default is 7.}

\item{dot.size.profile}{size of dots in profile plot. Default is 2.}
}
\description{
Create profile plot
}
\keyword{internal}
