% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.makeVolcano}
\alias{.makeVolcano}
\title{Create a volcano plot}
\usage{
.makeVolcano(
  input,
  label_name,
  log_base_FC,
  log_base_pval,
  x.lim,
  ProteinName,
  dot.size,
  y.limdown,
  y.limup,
  text.size,
  FCcutoff,
  sig,
  x.axis.size,
  y.axis.size,
  legend.size,
  log_adjp
)
}
\arguments{
\item{input}{data.table}

\item{label_name}{contrast label}

\item{log_base_FC}{2 or 10}

\item{log_base_pval}{2 or 10}

\item{ProteinName}{for volcano plot only, whether display protein names or not. TRUE (default) means protein names, which are significant, are displayed next to the points. FALSE means no protein names are displayed.}

\item{dot.size}{size of dots in volcano plot and comparison plot. Default is 3.}

\item{text.size}{size of ProteinName label in the graph for Volcano Plot. Default is 4.}

\item{FCcutoff}{for volcano plot or heatmap, whether involve fold change cutoff or not. FALSE (default) means no fold change cutoff is applied for significance analysis. FCcutoff = specific value means specific fold change cutoff is applied.}

\item{sig}{FDR cutoff for the adjusted p-values in heatmap and volcano plot. level of significance for comparison plot. 100(1-sig)\% confidence interval will be drawn.  sig=0.05 is default.}

\item{x.axis.size}{size of axes labels, e.g. name of the comparisons in heatmap, and in comparison plot. Default is 10.}

\item{y.axis.size}{size of axes labels, e.g. name of targeted proteins in heatmap. Default is 10.}

\item{legend.size}{size of legend for color at the bottom of volcano plot.  Default is 7.}
}
\description{
Create a volcano plot
}
\keyword{internal}
