% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{.nicePrint}
\alias{.nicePrint}
\title{Print a table nicely}
\usage{
.nicePrint(string_vector)
}
\arguments{
\item{string_vector}{character}
}
\value{
character
}
\description{
Print a table nicely
}
\keyword{internal}
