% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.normalizeMedian}
\alias{.normalizeMedian}
\title{Median normalization}
\usage{
.normalizeMedian(input)
}
\arguments{
\item{input}{`data.table` in standard MSstats format}
}
\description{
Median normalization
}
\keyword{internal}
