% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.normalizeQuantile}
\alias{.normalizeQuantile}
\title{Quantile normalization based on the `preprocessCore` package}
\usage{
.normalizeQuantile(input)
}
\arguments{
\item{input}{`data.table` in MSstats standard format}
}
\description{
Quantile normalization based on the `preprocessCore` package
}
\keyword{internal}
