% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstats-package.R
\name{.nullAppender}
\alias{.nullAppender}
\title{log4r appender used not to write messages}
\usage{
.nullAppender(level, ...)
}
\arguments{
\item{level}{log level}

\item{...}{messages - ignored}
}
\value{
NULL invisibly
}
\description{
A convenience function written to save time on checking if messages should
be printed or logs should be written to a file.
}
\keyword{internal}
