% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.prepareForDataProcess}
\alias{.prepareForDataProcess}
\title{Check validity of data already processed by MSstats converter}
\usage{
.prepareForDataProcess(input, ...)
}
\arguments{
\item{input}{data.frame of class `MSstatsValidated`}

\item{..}{additional parameters, currently ignored}
}
\description{
Check validity of data already processed by MSstats converter
}
\keyword{internal}
