% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.prepareForFeatureSelection}
\alias{.prepareForFeatureSelection}
\title{Re-format the data before feature selection}
\usage{
.prepareForFeatureSelection(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Re-format the data before feature selection
}
\keyword{internal}
