% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.prepareSingleProteinForGC}
\alias{.prepareSingleProteinForGC}
\title{Prepare data for a single protein for group comparison}
\usage{
.prepareSingleProteinForGC(single_protein)
}
\arguments{
\item{single_protein}{data.table}
}
\description{
Prepare data for a single protein for group comparison
}
\keyword{internal}
