% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.preProcessIntensities}
\alias{.preProcessIntensities}
\title{Create ABUNDANCE column and log-transform intensities}
\usage{
.preProcessIntensities(input, log_base)
}
\arguments{
\item{input}{data.table}

\item{log_base}{base of the logarithm}
}
\description{
Create ABUNDANCE column and log-transform intensities
}
\keyword{internal}
