% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.quantileNormalizationSingleLabel}
\alias{.quantileNormalizationSingleLabel}
\title{Quantile normalization for a single label}
\usage{
.quantileNormalizationSingleLabel(input, runs, label = "L")
}
\arguments{
\item{input}{`data.table` in MSstats standard format}

\item{runs}{run labels}

\item{label}{"L" for light isotopes, "H" for heavy isotopes}
}
\description{
Quantile normalization for a single label
}
\keyword{internal}
