% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.replaceZerosWithNA}
\alias{.replaceZerosWithNA}
\title{Utility function for normalization: replace 0s by NA}
\usage{
.replaceZerosWithNA(vec)
}
\arguments{
\item{vec}{vector}
}
\description{
Utility function for normalization: replace 0s by NA
}
\keyword{internal}
