% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.saveSessionInfo}
\alias{.saveSessionInfo}
\title{Save information about R session to sessionInfo.txt file.}
\usage{
.saveSessionInfo()
}
\description{
Save information about R session to sessionInfo.txt file.
}
\keyword{internal}
