% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots_common.R
\name{.saveTable}
\alias{.saveTable}
\title{Save a data table to a file}
\usage{
.saveTable(input, folder_path, file_name)
}
\arguments{
\item{input}{data.table}

\item{folder_path}{path to a folder (or "" for working directory)}

\item{file_name}{name of a file to save. If this file already exists,
an integer will be appended to this name}
}
\description{
Save a data table to a file
}
\keyword{internal}
