% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.selectHighQualityFeatures}
\alias{.selectHighQualityFeatures}
\title{Select features of high quality}
\usage{
.selectHighQualityFeatures(input, min_feature_count, method = "highQuality")
}
\arguments{
\item{input}{data.table}

\item{min_feature_count}{minimum number of quality features to consider}

\item{method}{"highQuality" (default, \code{MASS::rlm}/QR-based robust
fit) or "highQuality-lsqr" (matrix-free LSQR-based robust fit, see
\code{.fitHuberLSQR})}
}
\value{
data.table
}
\description{
Select features of high quality
}
\keyword{internal}
