% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.selectTopFeatures}
\alias{.selectTopFeatures}
\title{Select features with highest average abundance}
\usage{
.selectTopFeatures(input, top_n)
}
\arguments{
\item{input}{data.table}

\item{top_n}{number of top features to select}
}
\value{
data.table
}
\description{
Select features with highest average abundance
}
\keyword{internal}
