% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_linear.R
\name{.summarizeLinear}
\alias{.summarizeLinear}
\title{Linear summarization}
\usage{
.summarizeLinear(input, cutoff_base, censored_symbol, remove50missing)
}
\arguments{
\item{input}{`data.table` processed by the `MSstatsSelectFeatures` function.}

\item{cutoff_base}{cutoff value for censoring (with censored_symbol = "NA"
or censored_symbol = "0"). "minFeature"/"minRun"/"minFeatureNRun".}

\item{censored_symbol}{Missing values are censored or at random. 
"NA" (default) assumes that all NAs in Intensity column are censored. 
"0" uses zero intensities as censored intensity. In this case, NA intensities 
are missing at random. The output from Skyline should use "0". 
Null assumes that all NA intensites are randomly missing.}

\item{remove50missing}{only for summaryMethod = "TMP". TRUE removes the runs 
which have more than 50% missing values. FALSE is default.}
}
\value{
data.table
}
\description{
Linear summarization
}
\keyword{internal}
