% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_linear.R
\name{.summarizeLinearSingleProtein}
\alias{.summarizeLinearSingleProtein}
\title{Linear summarization for a single protein}
\usage{
.summarizeLinearSingleProtein(input, equal_variances = TRUE)
}
\arguments{
\item{input}{data.table of feature-level data for a single protein}

\item{equal_variances}{if FALSE, fitted model will be adjusted 
for non-homogenous variance. Default is TRUE}
}
\value{
data.table
}
\description{
Linear summarization for a single protein
}
\keyword{internal}
