% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_tmp.R
\name{.summarizeTukey}
\alias{.summarizeTukey}
\title{Summarization with Tukey median polish}
\usage{
.summarizeTukey(
  input,
  impute,
  cutoff_base,
  censored_symbol,
  remove50missing,
  original_scale = FALSE,
  n_threads = NULL
)
}
\description{
Summarization with Tukey median polish
}
