% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_tmp.R
\name{.summarizeTukeySingleProtein}
\alias{.summarizeTukeySingleProtein}
\title{Tukey median polish for a single protein}
\usage{
.summarizeTukeySingleProtein(
  input,
  impute,
  cutoff_base,
  censored_symbol,
  remove50missing,
  n_threads = NULL
)
}
\arguments{
\item{input}{`data.table` processed by the `MSstatsSelectFeatures` function.}

\item{impute}{only for summaryMethod = "TMP" and censored_symbol = 'NA' or '0'. 
TRUE (default) imputes "NA" or "0" (depending on censored_symbol option) by 
Accelated failure model. FALSE uses the values assigned by cutoffCensored.}

\item{cutoff_base}{cutoff value for censoring (with censored_symbol = "NA"
or censored_symbol = "0").}

\item{censored_symbol}{Missing values are censored or at random. 
"NA" (default) assumes that all NAs in Intensity column are censored. 
"0" uses zero intensities as censored intensity. In this case, NA intensities 
are missing at random. The output from Skyline should use "0". 
Null assumes that all NA intensites are randomly missing.}

\item{remove50missing}{only for summaryMethod = "TMP". TRUE removes the runs 
which have more than 50\% missing values. FALSE is default.}

\item{n_threads}{currently ignored. In the future, number of threads used
for parallel processing.}
}
\value{
data.table
}
\description{
Tukey median polish for a single protein
}
\keyword{internal}
