% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.updateColumnsForProcessing}
\alias{.updateColumnsForProcessing}
\title{Create columns for data processing}
\usage{
.updateColumnsForProcessing(input)
}
\arguments{
\item{input}{data.table}
}
\description{
Create columns for data processing
}
\keyword{internal}
