% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.updateUnequalVariances}
\alias{.updateUnequalVariances}
\title{Adjust model for unequal variances}
\usage{
.updateUnequalVariances(input, fit, num_iter)
}
\arguments{
\item{input}{data.table}

\item{fit}{lm}

\item{num_iter}{number of iterations}
}
\value{
merMod
}
\description{
Adjust model for unequal variances
}
\keyword{internal}
