\name{example_SDRF}
\alias{example_SDRF}
\title{Example SDRF.}
\description{An example SDRF file which is used to store metadata for MS-based protemics experiments.}

\usage{example_SDRF}

\format{data.frame}

\details{An example SDRF file which is used to store metadata for MS-based protemics experiments.
}

\value{data.frame example of an SDRF file.}

\author{Mateusz Staniak, Devon Kohler, Olga Vitek. 
}

\examples{
head(example_SDRF)
}

