% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{extract_multifrac_data}
\alias{extract_multifrac_data}
\title{Examine the work data structure for multi fraction data.}
\usage{
extract_multifrac_data(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
