% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{extract_singlefrac_labeled_data}
\alias{extract_singlefrac_labeled_data}
\title{Examine the work data structure single fraction labeled data.}
\usage{
extract_singlefrac_labeled_data(work, fillIncompleteRows = TRUE)
}
\arguments{
\item{work}{The work data structure from dataProcess()}

\item{fillIncompleteRows}{Fill those rows if they are not finished.}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
