% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{fit.quantification.run}
\alias{fit.quantification.run}
\title{Use lm to perform the actual quantification}
\usage{
fit.quantification.run(sub, singleFeature, singleSubject, TechReplicate,
  labeled, equalFeatureVar)
}
\arguments{
\item{sub}{yeah I dunnot yet, I need to read the context of this function.}

\item{singleFeature}{A single feature, I am guessing.}

\item{singleSubject}{and a subject}

\item{TechReplicate}{look, if you are going to do camelCase, then why
CamelCase?  I bet you have some smart reason, but it seems like fuckery to me.}

\item{labeled}{whatever}

\item{equalFeatureVar}{set var for the equal features?}
}
\value{
a fit from lm
}
\description{
Use lm to perform the actual quantification
}
