% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{getSamplesInfo}
\alias{getSamplesInfo}
\title{Get information about number of measurements for each group}
\usage{
getSamplesInfo(summarization_output)
}
\arguments{
\item{summarization_output}{output of the dataProcess function}
}
\value{
data.table
}
\description{
Get information about number of measurements for each group
}
