% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots_common.R
\name{getSelectedProteins}
\alias{getSelectedProteins}
\title{Get proteins based on names or integer IDs}
\usage{
getSelectedProteins(chosen_proteins, all_proteins)
}
\arguments{
\item{chosen_proteins}{protein names or integers IDs}

\item{all_proteins}{all unique proteins}
}
\value{
character
}
\description{
Get proteins based on names or integer IDs
}
