% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubnetworkFromIndra.R
\name{getSubnetworkFromIndra}
\alias{getSubnetworkFromIndra}
\title{Get subnetwork from INDRA database with differential analysis results}
\usage{
getSubnetworkFromIndra(input, pvalue_cutoff = NULL)
}
\arguments{
\item{input}{groupComparison comparisionResult table with additional HGNC ID 
and HGNC name columns}

\item{pvalue_cutoff}{p-value cutoff for filtering. Default is NULL, i.e. no 
filtering}
}
\value{
list of 2 data.frames, nodes and edges
}
\description{
Get subnetwork from INDRA database with differential analysis results
}
\examples{
input = data.table::fread(system.file("tinytest/processed_data/groupComparisonModel.csv", 
                             package = "MSstats"))
# subnetwork = getSubnetworkFromIndra(input, pvalue_cutoff = 0.05)
# head(subnetwork$nodes)
# head(subnetwork$edges)

  
}
