% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_converters.R
\name{MaxQtoMSstatsFormat}
\alias{MaxQtoMSstatsFormat}
\title{Import MaxQuant files}
\usage{
MaxQtoMSstatsFormat(
  evidence,
  annotation,
  proteinGroups,
  proteinID = "Proteins",
  useUniquePeptide = TRUE,
  summaryforMultipleRows = max,
  fewMeasurements = "remove",
  removeMpeptides = FALSE,
  removeOxidationMpeptides = FALSE,
  removeProtein_with1Peptide = FALSE,
  ...
)
}
\arguments{
\item{evidence}{name of 'evidence.txt' data, which includes feature-level data.}

\item{annotation}{name of 'annotation.txt' data which includes Raw.file, Condition, BioReplicate, Run, IsotopeLabelType information.}

\item{proteinGroups}{name of 'proteinGroups.txt' data. It needs to matching protein group ID. If proteinGroups=NULL, use 'Proteins' column in 'evidence.txt'.}

\item{proteinID}{'Proteins'(default) or 'Leading.razor.protein' for Protein ID.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import MaxQuant files
}
\note{
Warning: MSstats does not support for metabolic labeling or iTRAQ experiments.
}
\author{
Meena Choi, Olga Vitek.
}
