% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats.R
\docType{package}
\name{MSstats}
\alias{MSstats}
\alias{MSstats-package}
\title{MSstats: a suite of tools to make MS analyses easier.}
\description{
This provides a series of helpers for working with MS/MS data, DDA/DIA/etc.
}
