% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparison.R
\name{MSstatsPrepareForGroupComparison}
\alias{MSstatsPrepareForGroupComparison}
\title{Prepare output for dataProcess for group comparison}
\usage{
MSstatsPrepareForGroupComparison(summarization_output)
}
\arguments{
\item{summarization_output}{output of dataProcess}
}
\value{
list of run-level data for each protein in the input.
This list has a "has_imputed" attribute that indicates if missing values
were imputed in the input dataset.
}
\description{
Prepare output for dataProcess for group comparison
}
\examples{
QuantData <- dataProcess(SRMRawData, use_log_file = FALSE)
group_comparison_input <- MSstatsPrepareForGroupComparison(QuantData)
length(group_comparison_input) # list of length equal to number of proteins
# in protein-level data of QuantData
head(group_comparison_input[[1]])
}
