% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{multifraction_normalize_confusing}
\alias{multifraction_normalize_confusing}
\title{Examine the work data structure and normalize multifraction data
This code is confusering.}
\usage{
multifraction_normalize_confusing(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
