% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_quantlim.R
\name{nonlinear_quantlim}
\alias{nonlinear_quantlim}
\title{I have read not one word of this function yet, I will figure out what it does now.}
\usage{
nonlinear_quantlim(datain, alpha = 0.05, Npoints = 100,
  do_bootstrap = TRUE, Nbootstrap = 500)
}
\arguments{
\item{datain}{Well, I would guess the input data, presumably from dataProcess()}

\item{alpha}{Ah the first sign of code written by a smart person, a variable
named after a greek letter...}

\item{Npoints}{I am going to guess the number of points, points of what I
have no clue.}

\item{do_bootstrap}{Yes!}

\item{Nbootstrap}{The number bootstrap replicates I presume.}
}
\description{
I have read not one word of this function yet, I will figure out what it does now.
}
