% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{normalize_by_globalstandards}
\alias{normalize_by_globalstandards}
\title{Examine the work data structure and normalize by global standards.}
\usage{
normalize_by_globalstandards(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
