% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{normalize_singlefrac_by_heavy_standard}
\alias{normalize_singlefrac_by_heavy_standard}
\title{Examine the work data structure and normalize single fraction data by heavy standards.}
\usage{
normalize_singlefrac_by_heavy_standard(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
