% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats.R
\name{.onLoad}
\alias{.onLoad}
\title{Set up a log4j-ish log.}
\usage{
.onLoad(libname, pkgname)
}
\description{
There are a few places in this code which do various logging tasks.
There exist a few packages in R which handle this type of work reasonably well,
I arbitrarily chose 'logging'.
}
\keyword{internal}
