% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{OpenMStoMSstatsFormat}
\alias{OpenMStoMSstatsFormat}
\title{Import OpenMS files}
\usage{
OpenMStoMSstatsFormat(
  input,
  annotation = NULL,
  useUniquePeptide = TRUE,
  fewMeasurements = "remove",
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  ...
)
}
\arguments{
\item{input}{name of MSstats input report from OpenMS, which includes feature(peptide ion)-level data.}

\item{annotation}{name of 'annotation.txt' data which includes Condition, BioReplicate, Run. 
Run should be the same as filename.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import OpenMS files
}
\author{
Meena Choi, Olga Vitek.
}
