% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{OpenSWATHtoMSstatsFormat}
\alias{OpenSWATHtoMSstatsFormat}
\title{Import OpenSWATH files}
\usage{
OpenSWATHtoMSstatsFormat(
  input,
  annotation,
  filter_with_mscore = TRUE,
  mscore_cutoff = 0.01,
  useUniquePeptide = TRUE,
  fewMeasurements = "remove",
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  ...
)
}
\arguments{
\item{input}{name of MSstats input report from OpenSWATH, which includes feature-level data.}

\item{annotation}{name of 'annotation.txt' data which includes Condition, BioReplicate, Run. 
Run should be the same as filename.}

\item{filter_with_mscore}{TRUE(default) will filter out the features that have greater than mscore_cutoff in m_score column. Those features will be removed.}

\item{mscore_cutoff}{Cutoff for m_score. Default is 0.01.}

\item{useUniquePeptide}{TRUE (default) removes peptides that are assigned for more than one proteins. 
We assume to use unique peptide for each protein.}

\item{fewMeasurements}{'remove'(default) will remove the features that have 1 or 2 measurements across runs.}

\item{removeProtein_with1Feature}{TRUE will remove the proteins which have only 1 feature, which is the combination of peptide, precursor charge, fragment and charge. FALSE is default.}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple measurements for certain feature and certain run, use highest or sum of multiple intensities.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import OpenSWATH files
}
\author{
Meena Choi, Olga Vitek.
}
