% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDtoMSstatsFormat.R
\name{PDtoMSstatsFormat}
\alias{PDtoMSstatsFormat}
\title{Converter for Proteome discoverer PSM sheet to MSstats.}
\usage{
PDtoMSstatsFormat(input, annotation, useNumProteinsColumn = FALSE,
  useUniquePeptide = TRUE, summaryforMultipleRows = max,
  fewMeasurements = "remove", removeOxidationMpeptides = FALSE,
  removeProtein_with1Peptide = FALSE,
  which.quantification = "Precursor.Area",
  which.proteinid = "Protein.Group.Accessions", which.sequence = "Sequence")
}
\arguments{
\item{input}{Proteome discoverer PSM sheet to convert}

\item{annotation}{Dataframe of peptide/protein annotations.}

\item{useNumProteinsColumn}{What it says on the tin?  Though I am curious
why one would not want that information, is it not useful?}

\item{useUniquePeptide}{Instead of Proteins, use the Unique Peptides?}

\item{summaryforMultipleRows}{Presumably this is the function passed to
aggregate, but I haven't read this code carefully yet.}

\item{fewMeasurements}{What to do with the drunken sailor^H^H^H the peptides
with few measurements.}

\item{removeOxidationMpeptides}{This is another parameter which leaves me
thinking that there is a reason one would want to remove oxidation
products, but I do not know what it is.}

\item{removeProtein_with1Peptide}{Does what it says on the poorly labeled
tin.}

\item{which.quantification}{Which column to use for quantification?}

\item{which.proteinid}{Choose a column for gathering protein IDs.}

\item{which.sequence}{Which column contains the peptide sequence?}
}
\value{
A dataframe suitable for dataProcess()
}
\description{
Stupid question: Is proteome discoverer the thermo fisher software?  I have
only seen our MSMS person using it, but it looks eminently hate-able.
}
