% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{perform_survival_censored}
\alias{perform_survival_censored}
\title{Use Rs Survival package to quantify the censored data.}
\usage{
perform_survival_censored(modified_data, equalFeatureVar = TRUE,
  cutoffCensored = "minFeature", censoredInt = "NA",
  remove50missing = FALSE, MBimpute = TRUE, logsum = FALSE,
  featureSubset = "all", message.show = FALSE, clusters = 1,
  original_scale = FALSE, pct_adjust = 0.99, ...)
}
\arguments{
\item{modified_data}{the MS dataset after filtering and such.}

\item{equalFeatureVar}{I think this variable is not used, it is top of my
list to remove.}

\item{cutoffCensored}{What criterion for censoring?}

\item{censoredInt}{"NA" not NA!}

\item{remove50missing}{Drop proteins for which we have < 50% coverage (I
think?)}

\item{MBimpute}{Impute Impute damn spot!}

\item{logsum}{I am not sure.}

\item{featureSubset}{Keep them all!  I think this parameter is not used.}

\item{message.show}{Print only the duds which this is false.}

\item{clusters}{I think this is not used for this function.}

\item{original_scale}{Put the data back on the original scale?}

\item{pct_adjust}{Adjust the data by this factor?}

\item{...}{Extra arguments dropped into arglist for future reference.}
}
\value{
a dataframe of quantified proteins/peptides.
}
\description{
I have not yet carefully read this code yet, so I dunno what it does properly.
}
