% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{perform_tukey_polish_median}
\alias{perform_tukey_polish_median}
\title{I need to read up on this}
\usage{
perform_tukey_polish_median(modified_data, label = 1, pct_adjust = 0.99,
  clusters = 1, MBimpute = TRUE, censoredInt = "NA",
  message.show = FALSE, cutoffCensored = "minFeature",
  original_scale = FALSE, remove50missing = FALSE, ...)
}
\arguments{
\item{modified_data}{The data provided to msstats after some sanitization.}

\item{label}{Is this data labeled?}

\item{pct_adjust}{This is a peculiar feature of this code, this adjustment
is applied liberally throughout, and I am not sure yet which it is not done
just once at the beginning.  But until I figure that out I will not mess
with it.}

\item{clusters}{Currently not used until I clean it up enough that I feel
confident re-invoking parallel}

\item{MBimpute}{Perform imputation?}

\item{censoredInt}{"NA", not NA!}

\item{message.show}{Only print the peptides which fail when false.}

\item{cutoffCensored}{Where is the cutoff for censoring?}

\item{original_scale}{Keep the data on its original scale?}

\item{remove50missing}{If coverage is < 50%, drop the protein?}

\item{...}{I might add more stuff, so drop it into arglist.}
}
\value{
a list of tukeys!
}
\description{
I am familiar with some statistics tests from Tukey, but not the polish median.
I know for a fact that not all of these options are required, but I am not
ready to remove them yet.
}
