% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantlim.R
\name{plot_quantlim}
\alias{plot_quantlim}
\title{The goal of this function is to calculate the LOB/LOD of the data provided in the data frame.}
\usage{
plot_quantlim(spikeindata, quantlim_out, alpha, dir_output, xlim_plot)
}
\arguments{
\item{spikeindata}{Data from our spiked-in proteins}

\item{quantlim_out}{Output from the quantlim function}

\item{alpha}{damn smart people}

\item{dir_output}{Presumably a place to put the output}

\item{Plot}{the xlim!?}
}
\value{
something which I don't know what it is...
}
\description{
The function returns a new data frame containing the value of the LOB/LOD
}
