% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_converters.R
\name{ProgenesistoMSstatsFormat}
\alias{ProgenesistoMSstatsFormat}
\title{Import Progenesis files}
\usage{
ProgenesistoMSstatsFormat(
  input,
  annotation,
  useUniquePeptide = TRUE,
  summaryforMultipleRows = max,
  fewMeasurements = "remove",
  removeOxidationMpeptides = FALSE,
  removeProtein_with1Peptide = FALSE,
  ...
)
}
\arguments{
\item{input}{name of Progenesis output, which is wide-format. 'Accession', 'Sequence', 'Modification', 'Charge' and one column for each run are required.}

\item{annotation}{name of 'annotation.txt' or 'annotation.csv' data which includes Condition, BioReplicate, Run information. It will be matched with the column name of input for MS runs.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import Progenesis files
}
\author{
Meena Choi, Olga Vitek, Ulrich Omasits
}
