% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{resultsAsLists}
\alias{resultsAsLists}
\title{Manual function allowing foreach to return a list of multiple variables
Neat!}
\usage{
resultsAsLists(x, ...)
}
\arguments{
\item{x}{A foreach result to recast}

\item{...}{The rest of the stuff to recast, I am guessing.}
}
\value{
A list of variables rather than foreach's mess
}
\description{
Manual function allowing foreach to return a list of multiple variables
Neat!
}
