% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{runQuantification}
\alias{runQuantification}
\title{Perform the quantification for the sanitized data}
\usage{
runQuantification(data, summaryMethod = "TMP", equalFeatureVar = TRUE,
  cutoffCensored = "minFeature", censoredInt = "NA",
  remove50missing = FALSE, MBimpute = TRUE, logsum = FALSE,
  featureSubset = "all", message.show = FALSE, clusters = 1,
  original_scale = FALSE, pct_adjust = 0.99)
}
\arguments{
\item{data}{Yay!}

\item{summaryMethod}{Which method to use for summarizing}

\item{equalFeatureVar}{Set vars for equal features?}

\item{cutoffCensored}{Remove censored values?}

\item{censoredInt}{Int for censored!}

\item{remove50missing}{Yeah I hate those 50 missing guys, too.}

\item{MBimpute}{Impute Impute!}

\item{logsum}{oh yeah, or just don't do anything to separate words because
consistency is the hobgoblin of little minds.}

\item{featureSubset}{My eyes don't focus well enough to do camelCase
efficiently.  So I am sad that you came back this way.}

\item{message.show}{and I think using a '.' as a space is nasty.  ick.}

\item{clusters}{for my money, I would go with cluster_size, or cpus or
something.}

\item{original_scale}{ok, so this bugs me, camelCase, snake_case,
dotted.names, pick one to use most of the time!}
}
\value{
some new data?
}
\description{
I haven't read this function yet, so I have no clue what it does, that should be amended.
So for the moment I am going to put stupid strings into the parameter list
docstrings until I understand what is going on.  Holy crap, I thought
dataProcess() was long and obnoxious, this function is 1,500 lines of dense
stuff.  Sometimes I hate smart people, a stupider person would have split
this into smaller, understandable pieces.
}
