% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots_common.R
\name{savePlot}
\alias{savePlot}
\title{Save a plot to pdf file}
\usage{
savePlot(name_base, file_name, width, height)
}
\arguments{
\item{name_base}{path to a folder (or "" for working directory)}

\item{file_name}{name of a file to save. If this file already exists,
an integer will be appended to this name}

\item{width}{width of a plot}

\item{height}{height of a plot}
}
\description{
Save a plot to pdf file
}
