% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{SkylinetoMSstatsFormat}
\alias{SkylinetoMSstatsFormat}
\title{Import Skyline files}
\usage{
SkylinetoMSstatsFormat(
  input,
  annotation = NULL,
  removeiRT = TRUE,
  filter_with_Qvalue = TRUE,
  qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE,
  fewMeasurements = "remove",
  removeOxidationMpeptides = FALSE,
  removeProtein_with1Feature = FALSE,
  ...
)
}
\arguments{
\item{input}{name of MSstats input report from Skyline, which includes feature-level data.}

\item{annotation}{name of 'annotation.txt' data which includes Condition, BioReplicate, Run. If annotation is already complete in Skyline, use annotation=NULL (default). It will use the annotation information from input.}

\item{removeiRT}{TRUE (default) will remove the proteins or peptides which are labeld 'iRT' in 'StandardType' column. FALSE will keep them.}

\item{filter_with_Qvalue}{TRUE(default) will filter out the intensities that have greater than qvalue_cutoff in DetectionQValue column. Those intensities will be replaced with zero and will be considered as censored missing values for imputation purpose.}

\item{qvalue_cutoff}{Cutoff for DetectionQValue. default is 0.01.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import Skyline files
}
\author{
Meena Choi, Olga Vitek
}
