% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{subset_topn_result}
\alias{subset_topn_result}
\title{take the top n for each result.}
\usage{
subset_topn_result(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
