% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{summarize_tmp_mbimpute}
\alias{summarize_tmp_mbimpute}
\title{Examine the work data structure and summarize it by mbimpute.}
\usage{
summarize_tmp_mbimpute(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
