% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots_common.R
\name{theme_msstats}
\alias{theme_msstats}
\title{Theme for MSstats plots}
\usage{
theme_msstats(
  type,
  x.axis.size = 10,
  y.axis.size = 10,
  legend_size = 13,
  strip_background = element_rect(fill = "gray95"),
  strip_text_x = element_text(colour = c("black"), size = 14),
  legend_position = "top",
  legend_box = "vertical",
  text_angle = 0,
  text_hjust = NULL,
  text_vjust = NULL,
  ...
)
}
\arguments{
\item{type}{type of a plot}

\item{x.axis.size}{size of text on the x axis}

\item{y.axis.size}{size of text on the y axis}

\item{legend_size}{size of the legend}

\item{strip_background}{background of facet}

\item{strip_text_x}{size of text on facets}

\item{legend_position}{position of the legend}

\item{legend_box}{legend.box}

\item{text_angle}{angle of text on the x axis (for condition and comparison plots)}

\item{text_hjust}{hjust parameter for x axis text (for condition and comparison plots)}

\item{text_vjust}{vjust parameter for x axis text (for condition and comparison plots)}

\item{...}{additional parameters passed on to ggplot2::theme()}
}
\description{
Theme for MSstats plots
}
