% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformMSnSet.R
\name{transformMSstatsToMSnSet}
\alias{transformMSstatsToMSnSet}
\title{Generate a MSnSet from a MSstats data set.}
\usage{
transformMSstatsToMSnSet(data)
}
\arguments{
\item{data}{Input data from MSstats.}
}
\description{
Generate a MSnSet from a MSstats data set.
}
