% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{validateAnnotation}
\alias{validateAnnotation}
\title{Check if annotation matches intended experimental design}
\usage{
validateAnnotation(msstats_table, design_type = "group comparison")
}
\arguments{
\item{msstats_table}{output of a converter function}

\item{design_type}{character, "group comparison" or "repeated measures"}
}
\value{
TRUE if annotation file is consistent with intended experimental design. Otherwise, an error is thrown
}
\description{
Check if annotation matches intended experimental design
}
