% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlots.R
\name{visualizeNetworks}
\alias{visualizeNetworks}
\title{Create visualization of networks in plotly}
\usage{
visualizeNetworks(input)
}
\arguments{
\item{input}{list of Uniprot protein IDs}
}
\description{
Create visualization of networks in plotly
}
