% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSubnetwork.R
\name{visualizeSubnetwork}
\alias{visualizeSubnetwork}
\title{Create visualization of subnetwork in cytoscape}
\usage{
visualizeSubnetwork(nodes, edges)
}
\arguments{
\item{nodes}{dataframe of nodes}

\item{edges}{dataframe of edges}
}
\description{
Create visualization of subnetwork in cytoscape
}
\examples{
input = data.table::fread(system.file("tinytest/processed_data/groupComparisonModel.csv", 
                             package = "MSstats"))
# subnetwork = getSubnetworkFromIndra(input)
# visualizeSubnetwork(subnetwork$nodes, subnetwork$edges)

              
}
