% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{check_fractions}
\alias{check_fractions}
\title{Examine the work data structure for viable fractions.}
\usage{
check_fractions(work)
}
\arguments{
\item{work}{The work data structure from dataProcess()}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
