% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessPlots.R
\name{dataProcessPlots}
\alias{dataProcessPlots}
\title{Given a processedData() set, this function should provide a bunch of fun plots of it.}
\usage{
dataProcessPlots(data = data, type = type, featureName = "Transition",
  ylimUp = FALSE, ylimDown = FALSE, scale = FALSE, interval = "CI",
  x.axis.size = 10, y.axis.size = 10, text.size = 4, text.angle = 0,
  legend.size = 7, dot.size.profile = 2, dot.size.condition = 3,
  width = 10, height = 10, which.Protein = "all", originalPlot = TRUE,
  summaryPlot = TRUE, save_condition_plot_result = FALSE, address = "")
}
\arguments{
\item{data}{The result from processData()}

\item{type}{Which type of plot to produce? (ProfilePlot, QCPlot, or ConditionPlot)}

\item{featureName}{There are a few available features to plot, I know Transition is one.}

\item{ylimUp}{Provide a upper boundary}

\item{ylimDown}{or a lower boundary}

\item{scale}{Choose a scale for the plot?}

\item{interval}{Use confidence intervals?}

\item{x.axis.size}{Choose a size for the axes?}

\item{y.axis.size}{Also the y-axis.}

\item{text.size}{Choose a size for the text?}

\item{text.angle}{And an angle, stuff like this should be in arglist(...) IMO}

\item{legend.size}{Choose a legend size?}

\item{dot.size.profile}{How big of a glyph for the conditions?}

\item{width}{And a width for the plot}

\item{height}{and a height for it too!}

\item{which.Protein}{plot them all?}

\item{originalPlot}{Plot the original data?}

\item{summaryPlot}{and the summary of the data?}

\item{save_condition_plot_result}{ok, wtf?  We have camelCase, camel.dot.Case, dot.args, and now this too?}

\item{address}{Who wants it!?}
}
\value{
a plot!

I am thinking to split each of these into its own function rather than being
  forced to come here for plotting, then just have this function call the
  others.
}
\description{
Ideally, I would fill in starting here the various plots, but I am lazy and tired.
}
