% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizeCalculation.R
\name{designSampleSize}
\alias{designSampleSize}
\title{Perform a statistical power calculation given a dataset, experimental design,
and some parameters defining the desired parameters for analysis.}
\usage{
designSampleSize(data, desiredFC = 2, FDR = 0.05, numSample = TRUE,
  power = 0.9, labeled = FALSE, scopeofBioReplication = "expanded",
  interference = TRUE, equalFeatureVar = TRUE)
}
\arguments{
\item{data}{The dataset to examine, presumably after dataProcess()}

\item{desiredFC}{The fold-change used to define 'significant'}

\item{FDR}{The false discovery rate used to define 'significant'}

\item{numSample}{Use the number of samples in the data to inform the power
calculation?}

\item{power}{Floating point number to describe the desired power in the
data.  This is strange, as I only see it used in a if(isTRUE()), but it is
defined here as 0.9... hmmm I need to read this code.}

\item{scopeofBioReplication}{The following parameters were just defined right
after the function definition, so I put them into it, I dunno what they do
yet.}

\item{interference}{Personally, I hate it when people interfere with me, but
in this case, TRUE!}

\item{equalFeatureVar}{This is a parameter from dataProcess, I know that much.}
}
\value{
Dataframe describing the power in the data.
}
\description{
Perform a statistical power calculation given a dataset, experimental design,
and some parameters defining the desired parameters for analysis.
}
