% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizeCalculationClassification.R
\name{designSampleSizeClassificationPlots}
\alias{designSampleSizeClassificationPlots}
\title{Visualization for sample size calculation in classification problem}
\usage{
designSampleSizeClassificationPlots(data)
}
\arguments{
\item{data}{output from function \code{\link{designSampleSizeClassification}}}
}
\value{
Plot for sample size estimation. x-axis : sample size, y-axis: mean predictive accuracy. Color: different protein number.
}
\description{
To illustrate the mean classification accuracy under different protein number and sample size. The input is the result from function \code{\link{designSampleSizeClassification}}.
}
\details{
Data in the example is based on the results of sample size calculation in classification problem from function \code{\link{designSampleSizeClassification}}
}
\examples{
# Consider the training set from a colorectal cancer study
# Subjects are from control group or colorectal cancer group
# 72 proteins were targeted with SRM
require(MSstatsBioData)
set.seed(1235)
data(SRM_crc_training)
QuantCRCSRM <- dataProcess(SRM_crc_training, normalization = FALSE)
# estimate the mean predictive accuray under different sizes of training data
# n_sample is the number of different sample size to simulate
# Datasets with 10 different sample size and 3 different protein numbers are simulated 
result.crc.srm <- designSampleSizeClassification(data=QuantCRCSRM, 
n_sample = 10, 
sample_incr = 10, 
protein_desc = 0.33, 
iter = 50)
designSampleSizeClassificationPlots(data=result.crc.srm)
}
\references{
T. Huang et al.  TBD  2018
}
\author{
Ting Huang, Meena Choi, Olga Vitek. 

Maintainer: Meena Choi (\email{mnchoi67@gmail.com})
}
