% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizeCalculation.R
\name{designSampleSizePlots}
\alias{designSampleSizePlots}
\title{A couple plots generated when looking at sample sizes of a msstats dataset.}
\usage{
designSampleSizePlots(data, ...)
}
\arguments{
\item{data}{The dataProcess()'d data structure.}

\item{...}{Extra arguments for ggplot2.}
}
\value{
plots!
}
\description{
A couple plots generated when looking at sample sizes of a msstats dataset.
}
