% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{DIANNtoMSstatsFormat}
\alias{DIANNtoMSstatsFormat}
\title{Import Diann files}
\usage{
DIANNtoMSstatsFormat(
  input,
  annotation = NULL,
  global_qvalue_cutoff = 0.01,
  qvalue_cutoff = 0.01,
  pg_qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE,
  removeFewMeasurements = TRUE,
  removeOxidationMpeptides = TRUE,
  removeProtein_with1Feature = TRUE,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL,
  MBR = TRUE,
  ...
)
}
\arguments{
\item{input}{name of MSstats input report from Diann, which includes feature-level data.}

\item{annotation}{name of 'annotation.txt' data which includes Condition, BioReplicate, Run.}

\item{global_qvalue_cutoff}{The global qvalue cutoff}

\item{qvalue_cutoff}{local qvalue cutoff for library}

\item{pg_qvalue_cutoff}{local qvalue cutoff for protein groups Run should be the same as filename.}

\item{useUniquePeptide}{should unique pepties be removed}

\item{removeFewMeasurements}{should proteins with few measurements be removed}

\item{removeOxidationMpeptides}{should peptides with oxidation be removed}

\item{removeProtein_with1Feature}{should proteins with a single feature be removed}

\item{use_log_file}{logical. If TRUE, information about data processing
will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be added
to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing wil be printed
to the console.}

\item{log_file_path}{character. Path to a file to which information about 
data processing will be saved. 
If not provided, such a file will be created automatically.
If `append = TRUE`, has to be a valid path to a file.}

\item{MBR}{True if analysis was done with match between runs}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame in the MSstats required format.
}
\description{
Import Diann files
}
\examples{
\dontrun{
input = fread('diann_pooled_report.tsv')
annot = fread('Annotation.csv')
colnames(annot) = c('Condition', 'Run', 'BioReplicate')
input = DIANNtoMSstatsFormat(input, annotation = annot, MBR = F)
head(input)
}
}
\author{
Elijah Willie
}
