% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.addModelInformation}
\alias{.addModelInformation}
\title{Add model information}
\usage{
.addModelInformation(input)
}
\arguments{
\item{input}{data.table}
}
\value{
data.table
}
\description{
Add model information
}
\keyword{internal}
