% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSize.R
\name{.calculatePower}
\alias{.calculatePower}
\title{Power calculation}
\usage{
.calculatePower(desiredFC, median_sigma_error, numSample)
}
\arguments{
\item{desiredFC}{the range of a desired fold change which includes the lower 
and upper values of the desired fold change.}

\item{median_sigma_error}{median of error standard deviation}

\item{numSample}{minimal number of biological replicates per condition. 
TRUE represents you require to calculate the sample size for this category, 
else you should input the exact number of biological replicates.}
}
\description{
Power calculation
}
\keyword{internal}
