% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.calculateProteinVariance}
\alias{.calculateProteinVariance}
\title{Calculate protein variances}
\usage{
.calculateProteinVariance(input)
}
\arguments{
\item{input}{data.table}
}
\value{
list of residuals, degress of freedom and variances
}
\description{
Calculate protein variances
}
\keyword{internal}
