% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.checkDataProcessParams}
\alias{.checkDataProcessParams}
\title{Check validity of parameters to dataProcess function}
\usage{
.checkDataProcessParams(
  log_base,
  normalization_method,
  standards_names,
  address,
  fill_rows,
  feature_selection,
  summarization,
  imputation,
  n_clusters
)
}
\arguments{
\item{log_base}{of logarithmic transformation}

\item{normalization_method}{string: "quantile", "equalizemedians", "FALSE",
"NONE" or "globalStandards"}

\item{address}{string}

\item{fill_rows}{logical, if TRUE, missing run observations for each feature
will be added with INTENSITY = NA}

\item{feature_selection}{list with elements: remove_uninformative}

\item{summarization}{list with elements: method.}

\item{imputation}{list with elements: cutoff, symbol.}

\item{n_clusters}{integer}
}
\description{
Check validity of parameters to dataProcess function
}
\keyword{internal}
