% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{.checkDataProcessParams}
\alias{.checkDataProcessParams}
\title{Check validity of parameters to dataProcess function}
\usage{
.checkDataProcessParams(
  log_base,
  normalization_method,
  standards_names,
  feature_selection,
  summarization,
  imputation
)
}
\arguments{
\item{log_base}{of logarithmic transformation}

\item{normalization_method}{string: "quantile", "equalizemedians", "FALSE",
"NONE" or "globalStandards"}

\item{feature_selection}{list with elements: remove_uninformative}

\item{summarization}{list with elements: method.}

\item{imputation}{list with elements: cutoff, symbol.}
}
\description{
Check validity of parameters to dataProcess function
}
\keyword{internal}
