% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.countInformative}
\alias{.countInformative}
\title{Count informative features}
\usage{
.countInformative(input, column)
}
\arguments{
\item{input}{data.table}

\item{column}{name of a column used for filtering}
}
\value{
numeric
}
\description{
Count informative features
}
\keyword{internal}
