% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_converters.R
\name{.documentFunction}
\alias{.documentFunction}
\title{Import converters from MSstatsConvert package}
\usage{
.documentFunction(
  fewMeasurements,
  useUniquePeptide,
  summaryforMultipleRows,
  removeProtein_with1Feature,
  removeProtein_with1Protein,
  removeOxidationMpeptides,
  removeMpeptides
)
}
\description{
Import converters from MSstatsConvert package
}
