% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{.finalizeTMP}
\alias{.finalizeTMP}
\title{Summary statistics for output of TMP-based summarization}
\usage{
.finalizeTMP(input, censored_symbol, impute, predicted_survival)
}
\arguments{
\item{input}{feature-level data}

\item{censored_symbol}{censored missing value indicator}

\item{impute}{if TRUE, censored missing values were imputed}

\item{predicted_survival}{data.table of predicted survival values used for imputation}
}
\description{
Summary statistics for output of TMP-based summarization
}
\keyword{internal}
