% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.fitModelSingleProtein}
\alias{.fitModelSingleProtein}
\title{Fit model and perform group comparison for a single protein}
\usage{
.fitModelSingleProtein(
  input,
  contrast_matrix,
  has_tech_replicates,
  is_single_subject,
  repeated,
  groups,
  samples_info,
  save_fitted_models,
  has_imputed
)
}
\arguments{
\item{input}{data.table of summarized data}

\item{contrast_matrix}{contrast matrix}

\item{has_tech_replicates}{if TRUE, there are technical replicates}

\item{is_single_subject}{if TRUE, experiment consists of a single subject}

\item{repeated}{if TRUE, experiment consists of repeated measurements}

\item{groups}{unique labels for experimental conditions}

\item{samples_info}{number of runs per group}

\item{save_fitted_models}{if TRUE, fitted model will be saved. If FALSE,
it will be replaced by NULL}

\item{has_imputed}{if TRUE, missing values have been imputed by dataProcess}
}
\description{
Fit model and perform group comparison for a single protein
}
\keyword{internal}
