% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.fitTukey}
\alias{.fitTukey}
\title{Fit tukey median polish for a data matrix}
\usage{
.fitTukey(input)
}
\arguments{
\item{input}{data.table with data for a single protein}
}
\value{
data.table
}
\description{
Fit tukey median polish for a data matrix
}
\keyword{internal}
