% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{.flagUninformativeSingleLabel}
\alias{.flagUninformativeSingleLabel}
\title{Flag uninformative features}
\usage{
.flagUninformativeSingleLabel(input, min_feature_count = 2)
}
\arguments{
\item{input}{data.table}

\item{min_feature_count}{minimum number of quality features to consider}
}
\value{
data.table
}
\description{
Flag uninformative features
}
\keyword{internal}
