% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.getContrast}
\alias{.getContrast}
\title{Create a contrast for a model with only group as a fixed effect}
\usage{
.getContrast(input, contrast, coefs, groups)
}
\arguments{
\item{input}{summarized data for a single protein}

\item{coefs}{coefficients of a linear model (named vector)}

\item{groups}{unique group labels}

\item{contrast_matrix}{row of a contrast_matrix}
}
\description{
Create a contrast for a model with only group as a fixed effect
}
\keyword{internal}
