% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_tmp.R
\name{.getNonMissingFilterStats}
\alias{.getNonMissingFilterStats}
\title{Get a logical vector for non-missing values to calculate summary statistics}
\usage{
.getNonMissingFilterStats(input, censored_symbol)
}
\arguments{
\item{input}{data.table with data for a single protein}
}
\value{
data.table
}
\description{
Get a logical vector for non-missing values to calculate summary statistics
}
\keyword{internal}
