% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.getNonMissingFilterStats}
\alias{.getNonMissingFilterStats}
\title{Get a logical vector for non-missing values to calculate summary statistics}
\usage{
.getNonMissingFilterStats(input, censored_symbol)
}
\arguments{
\item{input}{data.table with data for a single protein}

\item{censored_symbol}{Missing values are censored or at random. 
'NA' (default) assumes that all 'NA's in 'Intensity' column are censored. 
'0' uses zero intensities as censored intensity. 
In this case, NA intensities are missing at random. 
The output from Skyline should use '0'. 
Null assumes that all NA intensites are randomly missing.}
}
\value{
data.table
}
\description{
Get a logical vector for non-missing values to calculate summary statistics
}
\keyword{internal}
