% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{.groupComparisonWithSingleCore}
\alias{.groupComparisonWithSingleCore}
\title{Perform group comparison per protein iteratively with a single loop}
\usage{
.groupComparisonWithSingleCore(
  summarized_list,
  contrast_matrix,
  save_fitted_models,
  repeated,
  samples_info
)
}
\arguments{
\item{summarized_list}{output of MSstatsPrepareForGroupComparison}

\item{contrast_matrix}{contrast matrix}

\item{save_fitted_models}{if TRUE, fitted models will be included in the output}

\item{repeated}{logical, output of checkRepeatedDesign function}

\item{samples_info}{data.table, output of getSamplesInfo function}
}
\description{
Perform group comparison per protein iteratively with a single loop
}
\keyword{internal}
