% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{.isSummarizable}
\alias{.isSummarizable}
\title{Check if a protein can be summarized with TMP}
\usage{
.isSummarizable(input, remove50missing)
}
\arguments{
\item{input}{data.table}

\item{remove50missing}{if TRUE, proteins with more than 50% missing values
in all runs will not be summarized}
}
\value{
data.table
}
\description{
Check if a protein can be summarized with TMP
}
\keyword{internal}
