% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_plots.R
\name{.makeComparison}
\alias{.makeComparison}
\title{Create comparison plot}
\usage{
.makeComparison(
  input,
  log_base,
  dot.size,
  x.axis.size,
  y.axis.size,
  text.angle,
  hjust,
  vjust,
  y.limdown,
  y.limup
)
}
\arguments{
\item{input}{data.table}

\item{log_base}{2 or 10}

\item{dot.size}{size of dots in volcano plot and comparison plot. Default is 3.}

\item{x.axis.size}{size of axes labels, e.g. name of the comparisons in heatmap, and in comparison plot. Default is 10.}

\item{y.axis.size}{size of axes labels, e.g. name of targeted proteins in heatmap. Default is 10.}

\item{text.angle}{angle of x-axis labels represented each comparison at the bottom of graph in comparison plot. Default is 0.}
}
\description{
Create comparison plot
}
\keyword{internal}
