% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocess_plots.R
\name{.makeConditionPlot}
\alias{.makeConditionPlot}
\title{Make condition plot}
\usage{
.makeConditionPlot(
  input,
  scale,
  single_protein,
  y.limdown,
  y.limup,
  x.axis.size,
  y.axis.size,
  text.size,
  text.angle,
  legend.size,
  dot.size.condition,
  yaxis.name
)
}
\arguments{
\item{input}{data.table}

\item{scale}{for "ConditionPlot" only, FALSE(default) means each conditional
level is not scaled at x-axis according to its actual value (equal space at
x-axis). TRUE means each conditional level is scaled at x-axis according to
its actual value (unequal space at x-axis).}

\item{single_protein}{data.table}

\item{x.axis.size}{size of x-axis labeling for "Run" in Profile Plot and
QC Plot, and "Condition" in Condition Plot. Default is 10.}

\item{y.axis.size}{size of y-axis labels. Default is 10.}

\item{text.size}{size of labels represented each condition at the top of
graph in Profile Plot and QC plot. Default is 4.}

\item{text.angle}{angle of labels represented each condition at the top
of graph in Profile Plot and QC plot or x-axis labeling in Condition plot.
Default is 0.}

\item{legend.size}{size of feature legend (transition-level or peptide-level)
above graph in Profile Plot. Default is 7.}

\item{dot.size.condition}{size of dots in condition plot. Default is 3.}
}
\description{
Make condition plot
}
\keyword{internal}
