% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocess_plots.R
\name{.makeSummaryProfilePlot}
\alias{.makeSummaryProfilePlot}
\title{Make summary profile plot}
\usage{
.makeSummaryProfilePlot(
  input,
  is_censored,
  y.limdown,
  y.limup,
  x.axis.size,
  y.axis.size,
  text.size,
  text.angle,
  legend.size,
  dot.size.profile,
  cumGroupAxis,
  yaxis.name,
  lineNameAxis,
  groupNametemp
)
}
\arguments{
\item{input}{data.table}

\item{is_censored}{TRUE if censored values were imputed}

\item{x.axis.size}{size of x-axis labeling for "Run" in Profile Plot and 
QC Plot, and "Condition" in Condition Plot. Default is 10.}

\item{y.axis.size}{size of y-axis labels. Default is 10.}

\item{text.size}{size of labels represented each condition at the top of 
graph in Profile Plot and QC plot. Default is 4.}

\item{text.angle}{angle of labels represented each condition at the top
of graph in Profile Plot and QC plot or x-axis labeling in Condition plot. 
Default is 0.}

\item{legend.size}{size of feature legend (transition-level or peptide-level)
above graph in Profile Plot. Default is 7.}

\item{dot.size.profile}{size of dots in profile plot. Default is 2.}
}
\description{
Make summary profile plot
}
\keyword{internal}
