% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{.normalizeGlobalStandards}
\alias{.normalizeGlobalStandards}
\title{Normalization based on standards}
\usage{
.normalizeGlobalStandards(input, peptides_dict, standards)
}
\arguments{
\item{input}{data.table in MSstats format}

\item{peptides_dict}{`data.table` of names of peptides and their corresponding 
features.}

\item{standards}{character vector with names of standards, required if 
"GLOBALSTANDARDS" method was selected.}
}
\description{
Normalization based on standards
}
\keyword{internal}
