% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_prepare.R
\name{.prepareTMP}
\alias{.prepareTMP}
\title{Prepare feature-level data for TMP summarization}
\usage{
.prepareTMP(input, impute, censored_symbol)
}
\arguments{
\item{input}{data.table}

\item{impute}{logical}

\item{censored_symbol}{"0"/"NA"}
}
\value{
data.table
}
\description{
Prepare feature-level data for TMP summarization
}
\keyword{internal}
