% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_tmp.R
\name{.runTukey}
\alias{.runTukey}
\title{Fit Tukey median polish}
\usage{
.runTukey(
  input,
  is_labeled,
  censored_symbol,
  remove50missing,
  original_scale = FALSE,
  log_base = 2
)
}
\arguments{
\item{input}{data.table with data for a single protein}

\item{is_labeled}{logical, if TRUE, data is coming from an SRM experiment}

\item{censored_symbol}{Missing values are censored or at random. 
"NA" (default) assumes that all NAs in Intensity column are censored. 
"0" uses zero intensities as censored intensity. In this case, NA intensities 
are missing at random. The output from Skyline should use "0". 
Null assumes that all NA intensites are randomly missing.}

\item{remove50missing}{only for summaryMethod = "TMP". TRUE removes the runs 
which have more than 50% missing values. FALSE is default.}

\item{original_scale}{if TRUE, data will be summarized on the original scale.}

\item{log_base}{base of the logarithm function used for ABUNDANCE column}
}
\value{
data.table
}
\description{
Fit Tukey median polish
}
\keyword{internal}
