% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_tmp.R
\name{.runTukey}
\alias{.runTukey}
\title{Fit Tukey median polish}
\usage{
.runTukey(input, is_labeled, censored_symbol, remove50missing)
}
\arguments{
\item{input}{data.table with data for a single protein}

\item{is_labeled}{logical, if TRUE, data is coming from an SRM experiment}
}
\value{
data.table
}
\description{
Fit Tukey median polish
}
\keyword{internal}
