% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_censored.R
\name{.setCensoredByThreshold}
\alias{.setCensoredByThreshold}
\title{Set censored values based on minimum in run/feature/run or feature}
\usage{
.setCensoredByThreshold(input, censored_symbol, remove50missing)
}
\arguments{
\item{input}{`data.table` in MSstats format}

\item{censored_symbol}{censoredInt parameter to `dataProcess`}

\item{remove50missing}{if TRUE, features with at least 50% missing values
will be removed}
}
\description{
Set censored values based on minimum in run/feature/run or feature
}
\keyword{internal}
