% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_prepare.R
\name{getProcessed}
\alias{getProcessed}
\title{Get feature-level data to be used in the MSstatsSummarizationOutput function}
\usage{
getProcessed(input)
}
\arguments{
\item{input}{data.table processed by dataProcess subfunctions}
}
\value{
data.table processed by dataProcess subfunctions
}
\description{
Get feature-level data to be used in the MSstatsSummarizationOutput function
}
