% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison.R
\name{getSamplesInfo}
\alias{getSamplesInfo}
\title{Get information about number of measurements for each group}
\usage{
getSamplesInfo(summarization_output)
}
\arguments{
\item{summarization_output}{output of the dataProcess function}
}
\value{
data.table
}
\description{
Get information about number of measurements for each group
}
\details{
This function extracts information required to compute percentages
of missing and imputed values in group comparison.
}
\examples{
QuantData <- dataProcess(DDARawData, use_log_file = FALSE)
samples_info <- getSamplesInfo(QuantData)
samples_info

}
