% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupComparison.R
\name{groupComparison}
\alias{groupComparison}
\title{Perform group comparisons using lm/lme4}
\usage{
groupComparison(contrast.matrix = contrast.matrix, data = data)
}
\arguments{
\item{contrast.matrix}{A contrast matrix for comparing}

\item{data}{the result from processedData()}
}
\value{
a comparison

Looking at this, I find myself wondering what is to stop me from using limma,
  this is doing much the same thing.
}
\description{
Perform group comparisons using lm/lme4
}
