% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparison.R
\name{groupComparison}
\alias{groupComparison}
\title{Whole plot testing}
\usage{
groupComparison(
  contrast.matrix,
  data,
  save_fitted_models = TRUE,
  log_base = 2,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL
)
}
\arguments{
\item{contrast.matrix}{comparison between conditions of interests.}

\item{data}{name of the (output of dataProcess function) data set.}

\item{save_fitted_models}{logical, if TRUE, fitted models will be added to
the output.}

\item{log_base}{base of the logarithm used in dataProcess.}

\item{use_log_file}{logical. If TRUE, information about data processing
will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be added
to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing wil be printed
to the console.}

\item{log_file_path}{character. Path to a file to which information about 
data processing will be saved. 
If not provided, such a file will be created automatically.
If `append = TRUE`, has to be a valid path to a file.}
}
\value{
list that consists of the following elements: (TODO)
}
\description{
Whole plot testing
}
\details{
contrast.matrix : comparison of interest. Based on the levels of conditions, specify 1 or -1 to the conditions of interests and 0 otherwise. The levels of conditions are sorted alphabetically. Command levels(QuantData$FeatureLevelData$GROUP_ORIGINAL) can illustrate the actual order of the levels of conditions.
The underlying model fitting functions are lm and lmer for the fixed effects model and mixed effects model, respectively.
The input of this function is the quantitative data from function (dataProcess).
}
