% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparison.R
\name{groupComparison}
\alias{groupComparison}
\title{Whole plot testing}
\usage{
groupComparison(contrast.matrix, data, save_fitted_models = TRUE, log_base = 2)
}
\arguments{
\item{contrast.matrix}{comparison between conditions of interests.}

\item{data}{name of the (output of dataProcess function) data set.}

\item{save_fitted_models}{logical, if TRUE, fitted models will be added to
the output.}

\item{logBase}{base of the logarithm used in dataProcess.}
}
\value{
list that consists of the following elements: (TODO)
}
\description{
Whole plot testing
}
\details{
contrast.matrix : comparison of interest. Based on the levels of conditions, specify 1 or -1 to the conditions of interests and 0 otherwise. The levels of conditions are sorted alphabetically. Command levels(QuantData$ProcessedData$GROUP_ORIGINAL) can illustrate the actual order of the levels of conditions.
The underlying model fitting functions are lm and lmer for the fixed effects model and mixed effects model, respectively.
The input of this function is the quantitative data from function (dataProcess).
}
