% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{makePeptidesDictionary}
\alias{makePeptidesDictionary}
\title{Prepare a peptides dictionary for global standards normalization}
\usage{
makePeptidesDictionary(input, normalization)
}
\arguments{
\item{input}{`data.table` in MSstats standard format}

\item{normalization}{normalization method}
}
\description{
Prepare a peptides dictionary for global standards normalization
}
