% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupComparison.R
\name{modelBasedQCPlots}
\alias{modelBasedQCPlots}
\title{Plot some QC metrics given a statistical model}
\usage{
modelBasedQCPlots(data, type, axis.size = 10, dot.size = 3, text.size = 7,
  legend.size = 7, width = 10, height = 10, which.Protein = "all",
  address = "")
}
\arguments{
\item{data}{processedData()}

\item{type}{the type of plot}

\item{axis.size}{How big for the axes?}

\item{dot.size}{How big for the dots?}

\item{text.size}{How big for the text?}

\item{legend.size}{And the legend?}

\item{width}{How wide?}

\item{height}{and high?}

\item{which.Protein}{Include which proteins?}

\item{address}{That's it, I'm getting my tinfoil hat!}

\item{featureName}{Include the peptides on the plot?}

\item{feature.QQPlot}{How many qqs to make?}
}
\value{
plots!
}
\description{
The first thing I am going to do is rip out the pdf calls, if I want to put
my plots into a file, I can damn well do it myself.
}
