% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupcomparison_contrasts.R
\name{MSstatsContrastMatrix}
\alias{MSstatsContrastMatrix}
\title{Create a contrast matrix for groupComparison function}
\usage{
MSstatsContrastMatrix(contrasts, conditions, labels = NULL)
}
\arguments{
\item{conditions}{unique condition labels}

\item{labels}{labels for contrasts (row.names of the contrast matrix)}

\item{contrast}{list of lists. Each sub-list consists of two vectors that 
name conditions that will be compared. See the details section for more 
information}
}
\description{
Create a contrast matrix for groupComparison function
}
