% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparison.R
\name{MSstatsGroupComparison}
\alias{MSstatsGroupComparison}
\title{Group comparison}
\usage{
MSstatsGroupComparison(
  summarized_list,
  contrast_matrix,
  save_fitted_models,
  repeated,
  samples_info
)
}
\arguments{
\item{summarized_list}{output of MSstatsPrepareForGroupComparison}

\item{contrast_matrix}{contrast matrix}

\item{repeated}{logical, output of checkRepeatedDesign function}

\item{samples_info}{data.table, output of getSamplesInfo function}

\item{save_fitted_model}{if TRUE, fitted models will be included in the output}
}
\description{
Group comparison
}
