% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparison.R
\name{MSstatsGroupComparisonOutput}
\alias{MSstatsGroupComparisonOutput}
\title{Create output of group comparison based on results for individual proteins}
\usage{
MSstatsGroupComparisonOutput(input, summarization_output, log_base = 2)
}
\arguments{
\item{input}{output of MSstatsGroupComparison function}

\item{summarization_output}{output of dataProcess function}

\item{log_base}{base of the logarithm used in fold-change calculation}
}
\description{
Create output of group comparison based on results for individual proteins
}
