% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_censored.R
\name{MSstatsHandleMissing}
\alias{MSstatsHandleMissing}
\title{Handle censored missing values}
\usage{
MSstatsHandleMissing(
  input,
  summary_method,
  impute,
  missing_symbol,
  censored_cutoff
)
}
\arguments{
\item{input}{`data.table` in MSstats data format}

\item{summary_method}{summarization method (`summaryMethod` parameter to `dataProcess`)}

\item{impute}{if TRUE, missing values are supposed to be imputed 
(`MBimpute` parameter to `dataProcess`)}

\item{missing_symbol}{`censoredInt` parameter to `dataProcess`}

\item{censored_cutoff}{`cutoffCensored` parameter to `dataProcess`}
}
\description{
Handle censored missing values
}
