% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{MSstatsMergeFractions}
\alias{MSstatsMergeFractions}
\title{Re-format the data before feature selection}
\usage{
MSstatsMergeFractions(input)
}
\arguments{
\item{input}{`data.table` in MSstats format}
}
\description{
Re-format the data before feature selection
}
