% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{MSstatsPrepareForDataProcess}
\alias{MSstatsPrepareForDataProcess}
\title{Prepare data for processing by `dataProcess` function}
\usage{
MSstatsPrepareForDataProcess(input, log_base)
}
\arguments{
\item{input}{`data.table` in MSstats format}

\item{log_base}{base of the logarithm to transform intensities}
}
\description{
Prepare data for processing by `dataProcess` function
}
