% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparison.R
\name{MSstatsPrepareForGroupComparison}
\alias{MSstatsPrepareForGroupComparison}
\title{Prepare output for dataProcess for group comparison}
\usage{
MSstatsPrepareForGroupComparison(summarization_output)
}
\arguments{
\item{summarization_output}{output of dataProcess}
}
\description{
Prepare output for dataProcess for group comparison
}
