% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{MSstatsSelectFeatures}
\alias{MSstatsSelectFeatures}
\title{Feature selection before feature-level data summarization}
\usage{
MSstatsSelectFeatures(input, method, top_n = NULL, min_feature_count = NULL)
}
\description{
Feature selection before feature-level data summarization
}
