% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature_selection.R
\name{MSstatsSelectFeatures}
\alias{MSstatsSelectFeatures}
\title{Feature selection before feature-level data summarization}
\usage{
MSstatsSelectFeatures(input, method, top_n = 3, min_feature_count = 2)
}
\arguments{
\item{input}{data.table}

\item{method}{"all" / "highQuality", "topN"}

\item{top_n}{number of features to use for "topN" method}

\item{min_feature_count}{number of quality features for "highQuality" method}
}
\value{
data.table
}
\description{
Feature selection before feature-level data summarization
}
