% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{MSstatsSummarizationOutput}
\alias{MSstatsSummarizationOutput}
\title{Post-processing output from MSstats summarization}
\usage{
MSstatsSummarizationOutput(input, summarized, summary_method)
}
\arguments{
\item{input}{`data.table` in MSstats format}

\item{summarized}{output of the `MSstatsSummarize` function}

\item{summary_method}{name of the summarization method
(`summaryMethod` parameter to `dataProcess`)}
}
\value{
list that consists of the following elements:
\item{ProcessedData - feature-level data after processing,} 
\item{RunlevelData - run-level (summarized) data,}
\item{SummaryMethod (string) - name of summarization method that was used.}
}
\description{
Post-processing output from MSstats summarization
}
