% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{MSstatsSummarizationOutput}
\alias{MSstatsSummarizationOutput}
\title{Post-processing output from MSstats summarization}
\usage{
MSstatsSummarizationOutput(
  input,
  summarized,
  processed,
  method,
  impute,
  censored_symbol
)
}
\arguments{
\item{input}{`data.table` in MSstats format}

\item{summarized}{output of the `MSstatsSummarize` function}

\item{processed}{output of MSstatsSelectFeatures}

\item{method}{name of the summarization method
(`summaryMethod` parameter to `dataProcess`)}

\item{impute}{if TRUE, censored missing values were imputed
(`MBimpute` parameter to `dataProcess`)}

\item{censored_symbol}{censored missing value indicator 
(`censoredInt` parameter to `dataProcess`)}
}
\value{
list that consists of the following elements:
\itemize{
\item{FeatureLevelData}{ - feature-level data after processing} 
\item{ProteinLevelData}{ - protein-level (summarized) data}
\item{SummaryMethod}{ (string) - name of summarization method that was used}
}
}
\description{
Post-processing output from MSstats summarization
}
