% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{MSstatsSummarize}
\alias{MSstatsSummarize}
\title{Feature-level data summarization}
\usage{
MSstatsSummarize(
  proteins_list,
  method,
  impute,
  censored_symbol,
  remove50missing,
  equal_variance
)
}
\arguments{
\item{method}{summarization method: "linear" or "TMP"}

\item{impute}{only for summaryMethod = "TMP" and censoredInt = 'NA' or '0'. 
TRUE (default) imputes 'NA' or '0' (depending on censoredInt option) by Accelated failure model. 
FALSE uses the values assigned by cutoffCensored}

\item{censored_symbol}{Missing values are censored or at random. 
'NA' (default) assumes that all 'NA's in 'Intensity' column are censored. 
'0' uses zero intensities as censored intensity. 
In this case, NA intensities are missing at random. 
The output from Skyline should use '0'. 
Null assumes that all NA intensites are randomly missing.}

\item{remove50missing}{only for summaryMethod = "TMP". TRUE removes the runs 
which have more than 50\% missing values. FALSE is default.}

\item{equal_variance}{only for summaryMethod = "linear". Default is TRUE. 
Logical variable for whether the model should account for heterogeneous variation 
among intensities from different features. Default is TRUE, which assume equal
variance among intensities from features. FALSE means that we cannot assume 
equal variance among intensities from features, then we will account for
heterogeneous variation from different features.}

\item{input}{list of processed feature-level data}

\item{remove_uninformative_feature_outlier}{It only works after users used featureSubset = "highQuality" 
in dataProcess. TRUE allows to remove 1) the features are flagged in the column, 
feature_quality = "Uninformative" which are features with bad quality, 
2) outliers that are flagged in the column, is_outlier = TRUE, 
for run-level summarization. FALSE (default) uses all features and intensities 
for run-level summarization.}
}
\description{
Feature-level data summarization
}
