% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization.R
\name{MSstatsSummarizeSingleLinear}
\alias{MSstatsSummarizeSingleLinear}
\title{Linear model-based summarization for a single protein}
\usage{
MSstatsSummarizeSingleLinear(single_protein, equal_variances = TRUE)
}
\arguments{
\item{single_protein}{feature-level data for a single protein}

\item{equal_variances}{if TRUE, observation are assumed to be homoskedastic}
}
\value{
list with protein-level data
}
\description{
Linear model-based summarization for a single protein
}
