% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{MSstatsSummarizeSingleTMP}
\alias{MSstatsSummarizeSingleTMP}
\title{Tukey Median Polish summarization for a single protein}
\usage{
MSstatsSummarizeSingleTMP(
  single_protein,
  impute,
  censored_symbol,
  remove50missing,
  aft_iterations = 90,
  aft_solver = "cholesky"
)
}
\arguments{
\item{single_protein}{feature-level data for a single protein}

\item{impute}{only for method = "TMP"; imputes censored values via AFT model when TRUE}

\item{censored_symbol}{how censored values are encoded: 'NA', '0', or NULL for none}

\item{remove50missing}{only for method = "TMP"; drops proteins missing >=50\% per peptide in every run}

\item{aft_iterations}{number of iterations for AFT model fitting}
}
\value{
list of two data.tables: one with fitted survival model,
the other with protein-level data
}
\description{
Tukey Median Polish summarization for a single protein
}
\examples{
raw = DDARawData 
method = "TMP"
cens = "NA"
impute = TRUE 
# currently, MSstats only supports MBimpute = FALSE for linear summarization
MSstatsConvert::MSstatsLogsSettings(FALSE)
input = MSstatsPrepareForDataProcess(raw, 2, NULL)
input = MSstatsNormalize(input, "EQUALIZEMEDIANS")
input = MSstatsMergeFractions(input)
input = MSstatsHandleMissing(input, "TMP", TRUE, "NA", 0.999)
input = MSstatsSelectFeatures(input, "all")
input = MSstatsPrepareForSummarization(input, method, impute, cens, FALSE)
input_split = split(input, input$PROTEIN)
single_protein_summary = MSstatsSummarizeSingleTMP(input_split[[1]],
                                                   impute, cens, FALSE, 100)
head(single_protein_summary[[1]])

}
