% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{OpenSWATHtoMSstatsFormat}
\alias{OpenSWATHtoMSstatsFormat}
\title{Import OpenSWATH files}
\usage{
OpenSWATHtoMSstatsFormat(
  input,
  annotation,
  filter_with_mscore = TRUE,
  mscore_cutoff = 0.01,
  useUniquePeptide = TRUE,
  fewMeasurements = "remove",
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  ...
)
}
\arguments{
\item{input}{name of MSstats input report from OpenSWATH, which includes feature-level data.}

\item{annotation}{name of 'annotation.txt' data which includes Condition, BioReplicate, Run. 
Run should be the same as filename.}

\item{filter_with_mscore}{TRUE(default) will filter out the features that have greater than mscore_cutoff in m_score column. Those features will be removed.}

\item{mscore_cutoff}{Cutoff for m_score. Default is 0.01.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import OpenSWATH files
}
\author{
Meena Choi, Olga Vitek.
}
