% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_converters.R
\name{PDtoMSstatsFormat}
\alias{PDtoMSstatsFormat}
\title{Import Proteome Discoverer files}
\usage{
PDtoMSstatsFormat(
  input,
  annotation,
  useNumProteinsColumn = FALSE,
  useUniquePeptide = TRUE,
  summaryforMultipleRows = max,
  fewMeasurements = "remove",
  removeOxidationMpeptides = FALSE,
  removeProtein_with1Peptide = FALSE,
  which.quantification = "Precursor.Area",
  which.proteinid = "Protein.Group.Accessions",
  which.sequence = "Sequence",
  ...
)
}
\arguments{
\item{input}{PD report or a path to it.}

\item{annotation}{name of 'annotation.txt' or 'annotation.csv' data which includes Condition, BioReplicate, 
Run information. 'Run' will be matched with 'Spectrum.File'.}

\item{useNumProteinsColumn}{TRUE removes peptides which have more than 1 in # Proteins column of PD output.}

\item{which.quantification}{Use 'Precursor.Area'(default) column for quantified intensities. 'Intensity' or 'Area' can be used instead.}

\item{which.proteinid}{Use 'Protein.Accessions'(default) column for protein name. 'Master.Protein.Accessions' can be used instead.}

\item{which.sequence}{Use 'Sequence'(default) column for peptide sequence. 'Annotated.Sequence' can be used instead.}

\item{...}{additional parameters to `data.table::fread`.}
}
\value{
data.frame with the required format of MSstats.
}
\description{
Import Proteome Discoverer files
}
\author{
Meena Choi, Olga Vitek
}
