% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantification.R
\name{quantification}
\alias{quantification}
\title{My voice is my password, quantify me.}
\usage{
quantification(data, type = "Sample", format = "matrix")
}
\arguments{
\item{data}{Input data from processData()}

\item{type}{Quantify based on which column in the experimental design?}

\item{format}{Return data as a ... matrix?}
}
\value{
the quantified data!
}
\description{
No seriously, this function performs quantifications on a MSMS data set.
}
