% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{summarize_tmp_mbimpute}
\alias{summarize_tmp_mbimpute}
\title{Examine the work data structure and summarize it by mbimpute.}
\usage{
summarize_tmp_mbimpute(work, maxQuantileforCensored = 0.999,
  censoredInt = "NA")
}
\arguments{
\item{work}{The work data structure from dataProcess()}

\item{maxQuantileforCensored}{What is the censoring cutoff by quantile?}

\item{censoredInt}{"NA!" not NA?}
}
\value{
Send the work data back to dataProcess()
}
\description{
This was just kill/yanked from dataProcess(), hopefully it does not depend on
anything there.
}
