% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformMSnSet.R
\name{transformMSnSetToMSstats}
\alias{transformMSnSetToMSstats}
\title{Transform a MSNset into a MSstat data set}
\usage{
transformMSnSetToMSstats(ProteinName, PeptideSequence, PrecursorCharge,
  FragmentIon, ProductCharge, IsotopeLabelType, Bioreplicate, Run, Condition,
  data)
}
\arguments{
\item{ProteinName}{A protein name!}

\item{PeptideSequence}{And the sequence of a peptide!}

\item{PrecursorCharge}{The charge of the observed precursor.}

\item{FragmentIon}{What is the state of the observed fragment?}

\item{ProductCharge}{The charge of the observed product.}

\item{IsotopeLabelType}{How were they labeled?}

\item{Bioreplicate}{What is the replicate state?}

\item{Run}{Which run is this?}

\item{Condition}{What is the condition of this sample?}

\item{data}{The data set to convert.

It seems to me that data should be the first argument and then fill in the
  rest of the arguments with default values.}
}
\description{
As someone who is new to proteomics, I know basically nothing about the MSNset or MSnSet or
whatever it is.  I guess this is my chance to learn...
}
